+++ vscode/src/chat/chat-view/ChatController.ts
@@ -544,6 +544,10 @@ export class ChatController implements vscode.Disposable, vscode.WebviewViewProv
                 logger(message.filterLabel, message.message)
                 break
             }
+            case 'chat/upload-image': {
+                await this.chatBuilder.addImage(message.image)
+                break
+            }
         }
     }
 
@@ -1763,6 +1767,7 @@ export class ChatController implements vscode.Disposable, vscode.WebviewViewProv
             const params = {
                 model,
                 maxTokensToSample: contextWindow.output,
+                images: this.chatBuilder.getAndResetImages(),
             } as CompletionParameters
 
             // Set stream param only when the model is disabled for streaming.